Changed the method for extracting the media_id from the URL
Instead of splitting the URL and using the last series of digits as the
media_id, i am now using a regex to locate any string of numbers 5
digits or larger. I had to do this because CR's POP video URLs have the
media_id in the middle of the URL instead of at the end like all the
anime and drama videos.